+++ src/ConfigProvider.php
@@ -14,7 +14,6 @@ use FlorianWolters\Component\Util\Singleton\SingletonTrait;
  * @property-read string env
  * @property-read string serverUrl
  * @property-read string capability
- * @property-read string publishResults
  * @property-read string fixturesDir
  * @property-read string logsDir
  * @property-read string debug
@@ -105,7 +104,6 @@ class ConfigProvider
             'ENV',
             'SERVER_URL',
             'CAPABILITY',
-            'PUBLISH_RESULTS',
             'FIXTURES_DIR',
             'LOGS_DIR',
             'DEBUG',